refactor(server): merge progress_stages.py into protocol.py
`StageId` is a wire-typed StrEnum living next to `MessageId` already —
both are i18n-key/protocol-string registries that the renderer mirrors.
Fold the 50-line file into protocol.py; consumers (server.py /
engine_manager.py / ws_session.py) flip from `from progress_stages
import StageId` to `from protocol import StageId`.

Refs #78.
